refactor(Stack): use gap instead of margin
