--- packages/orbit-components/src/Stack/helpers/getDirectionSpacingTemplate.js.flow
@@ -1,4 +1,0 @@
-// @flow
-export type GetDirectionSpacingTemplate = (direction: ?string) => string | boolean;
-
-declare export default GetDirectionSpacingTemplate;
