+++ packages/orbit-components/src/Stack/helpers/getGap.js.flow
@@ -5,10 +5,10 @@ import type { Devices } from "../../utils/mediaQuery/consts";
 import type { Props, SmallMobile } from "..";
 import type { ThemeProps } from "../../defaultTheme";
 
-export type GetChildrenMargin = ({|
+export type GetGap = ({|
   viewport: Devices,
   index: number,
   devices: Devices[],
 |}) => ({| ...Props, ...SmallMobile, ...ThemeProps |}) => InterpolationBase;
 
-declare export default GetChildrenMargin;
+declare export default GetGap;
