--- packages/orbit-components/src/Stack/helpers/shouldUseFlex.js.flow
@@ -1,6 +1,0 @@
-// @flow
-import type { Props } from "..";
-
-export type ShouldUseFlex = Props => boolean;
-
-declare export default ShouldUseFlex;
